DeleteCommandParser: Reject delete commands with both index and phone
The current implementation of DeleteCommandParser allows both index and phone number parameters simultaneously if the index is specified before the phone number, ignoring the index parameter. However, if the phone number parameter is specified first, an error is thrown to signify an invalid command format.

This introduces ambiguity - delete commands should have either an index or a phone number specified, but not both. This must be made clear to the end user.

Let's make DeleteCommandParser throw an error if both index and phone numbers are provided, no matter their order.